requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
hl7==0.4.5
idna==3.11
joblib==1.5.3
numpy==1.26.4
pandas==2.2.2
prometheus_client
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.5.0
scipy==1.17.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
urllib3==2.6.3
